// example_apps/guru_faucet/src/services/flow/schema.d.ts
export interface paths {
  '/api/health_check': {
    /** Health */
    get: operations['health_api_health_check_get']
  }
  '/api/users': {
    /** Create User */
    post: operations['create_user_api_users_post']
  }
  '/api/login/{webapp_user_id}': {
    /** Login */
    get: operations['login_api_login__webapp_user_id__get']
  }
  '/api/refresh': {
    /** Refresh Token */
    get: operations['refresh_token_api_refresh_get']
  }
  '/api/arts': {
    /** Get Arts Sys */
    get: operations['get_arts_sys_api_arts_get']
  }
  '/api/art': {
    /** Create Art */
    post: operations['create_art_api_art_post']
  }
  '/api/art/{art_id}': {
    /** Get Art Id */
    get: operations['get_art_id_api_art__art_id__get']
    /** Update Art */
    put: operations['update_art_api_art__art_id__put']
    /** Delete External Worker */
    delete: operations['delete_external_worker_api_art__art_id__delete']
  }
  '/api/art/{art_id}/like': {
    /** Like Art */
    post: operations['like_art_api_art__art_id__like_post']
  }
  '/api/arts/{art_id}': {
    /** Get Arts Id */
    get: operations['get_arts_id_api_arts__art_id__get']
  }
  '/api/art_collection': {
    /** Get Collections */
    get: operations['get_collections_api_art_collection_get']
    /** Create Collection */
    post: operations['create_collection_api_art_collection_post']
  }
  '/api/art_collection/{collection_id}': {
    /** Get Collection */
    get: operations['get_collection_api_art_collection__collection_id__get']
    /** Update Collection */
    put: operations['update_collection_api_art_collection__collection_id__put']
    /** Delete Collection */
    delete: operations['delete_collection_api_art_collection__collection_id__delete']
  }
// ...
  '/api/external_workers': {
    /** Get External Workers */
    get: operations['get_external_workers_api_external_workers_get']
    /** Create External Worker */
    post: operations['create_external_worker_api_external_workers_post']
  }
  '/api/external_workers/{external_worker_id}': {
    /** Get External Worker */
    get: operations['get_external_worker_api_external_workers__external_worker_id__get']
    /** Update External Worker */
    put: operations['update_external_worker_api_external_workers__external_worker_id__put']
    /** Delete External Worker */
    delete: operations['delete_external_worker_api_external_workers__external_worker_id__delete']
  }
  '/api/strategy': {
    /** Get Strategies */
    get: operations['get_strategies_api_strategy_get']
    /** Create Strategy */
    post: operations['create_strategy_api_strategy_post']
  }
  '/api/flows': {
    /** Get Flows */
    get: operations['get_flows_api_flows_get']
  }
  '/api/flow/{flow_id}': {
    /** Get Flow */
    get: operations['get_flow_api_flow__flow_id__get']
    /** Update Flow */
    put: operations['update_flow_api_flow__flow_id__put']
    /** Delete Flow */
    delete: operations['delete_flow_api_flow__flow_id__delete']
  }
  '/api/flow': {
    /** Create Flow */
    post: operations['create_flow_api_flow_post']
  }
  '/api/strategy/{strategy_id}': {
    /** Get Strategy */
    get: operations['get_strategy_api_strategy__strategy_id__get']
    /** Update Strategy */
    put: operations['update_strategy_api_strategy__strategy_id__put']
    /** Delete Strategy */
    delete: operations['delete_strategy_api_strategy__strategy_id__delete']
  }
  '/api/invites/chain/{chain_id}': {
    /** Get Invited Wallets */
    get: operations['get_invited_wallets_api_invites_chain__chain_id__get']
  }
  '/api/invites/chain/{chain_id}/token/{token_id}': {
    /** Get Invited Wallets By Token Id */
    get: operations['get_invited_wallets_by_token_id_api_invites_chain__chain_id__token__token_id__get']
  }
  '/engine/process-definition': {
    /** Get Process Definitions */
    get: operations['get_process_definitions_engine_process_definition_get']
  }
  '/engine/process-definition/key/{key}/start': {
    /** Start Process Instance By Key */
    post: operations['start_process_instance_by_key_engine_process_definition_key__key__start_post']
  }
  '/engine/task': {
    /**
     * Get Tasks
     * @description Get User's Tasks
     */
    get: operations['get_tasks_engine_task_get']
    /**
     * Filter Tasks
     * @description Filter User's Task
     */
    post: operations['filter_tasks_engine_task_post']
  }
  '/engine/task/{task_id}/form-variables': {
    /** Get Form Variables */
    get: operations['get_form_variables_engine_task__task_id__form_variables_get']
  }
  '/engine/task/{task_id}/complete': {
    /** Complete Task */
    post: operations['complete_task_engine_task__task_id__complete_post']
  }
  '/engine/task/{task_id}/variables': {
    /** Modify Task Variables */
    post: operations['modify_task_variables_engine_task__task_id__variables_post']
  }
  '/engine/process-instance': {
    /** Get Process Instances */
    get: operations['get_process_instances_engine_process_instance_get']
  }
  '/engine/process-instance/count': {
    /** Get Process Instance Count */
    get: operations['get_process_instance_count_engine_process_instance_count_get']
  }
  '/engine/process-instance/{instance_id}': {
    /** Get Process Instance */
    get: operations['get_process_instance_engine_process_instance__instance_id__get']
  }
  '/engine/process-instance/{instance_id}/variables': {
    /** Get Process Instance Variables */
    get: operations['get_process_instance_variables_engine_process_instance__instance_id__variables_get']
  }
  '/engine/process-instance/{process_id}': {
    /** Delete Process Instance */
    delete: operations['delete_process_instance_engine_process_instance__process_id__delete']
  }
  '/engine/process-instance/{instance_id}/suspended': {
    /** Update Suspension State By Id */
    put: operations['update_suspension_state_by_id_engine_process_instance__instance_id__suspended_put']
  }
// ...
}

export type webhooks = Record<string, never>

export interface components {
  schemas: {
// ...
    Body_filter_tasks_engine_task_post: {
      /** @description Filter user's tasks */
      schema?: components['schemas']['GetTasksFilterSchema'] | null
      /** @description Pagination parameters */
      pagination?: components['schemas']['PaginationParams'] | null
    }
    /** Body_start_process_instance_by_key_engine_process_definition_key__key__start_post */
    Body_start_process_instance_by_key_engine_process_definition_key__key__start_post: {
      /** Variables */
      variables?: {
        [key: string]: components['schemas']['VariableValueSchema']
      }
      /**
       * Business Key
       * @description Business key
       */
      business_key?: string | null
    }
// ...
    VariableValueSchema: {
      /** Value */
      value: unknown
      /** Type */
      type?:
        | (
            | 'Boolean'
            | 'Bytes'
            | 'Short'
            | 'Integer'
            | 'Long'
            | 'Double'
            | 'Date'
            | 'String'
            | 'Null'
            | 'Object'
            | 'Json'
          )
        | null
      /** Valueinfo */
      valueInfo?: Record<string, never> | null
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
  /** Health */
  health_api_health_check_get: {
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          'application/json': unknown
        }
      }
    }
  }
  /** Create User */
  create_user_api_users_post: {
    requestBody: {
      content: {
        'application/json': components['schemas']['Body_create_user_api_users_post']
      }
    }
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          'application/json': components['schemas']['UserRest']
        }
      }
      /** @description Validation Error */
      422: {
        content: {
          'application/json': components['schemas']['HTTPValidationError']
        }
      }
    }
  }
// ...
}